pyproject.toml
[project]
name = "myaicode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "any-llm-sdk[anthropic,ollama]>=0.17.1",
    "pydantic-ai>=0.8.1",
    "pydantic-ai-slim[anthropic]>=0.8.1",
]
